requirements.txt
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
transformers
openai
tiktoken
pypdf
chromadb
python-dotenv
streamlit
fastapi
uvicorn
requests
pandas
numpy
